requirements.txt
colorama==0.4.6
dash==2.18.1
dash_bootstrap_components==1.6.0
dash_core_components==2.0.0
dash_draggable==0.1.2
dash_extensions==1.0.18
dash_html_components==2.0.0
dash_iconify==0.1.2
dash_leaflet==1.0.15
dash_mantine_components==0.14.4
exceptions==0.1.0
fastapi==0.115.0
fastapi_utils==0.7.0
Flask==3.0.3
geojson==3.1.0
matplotlib==3.9.2
numpy==2.1.1
paho_mqtt==2.1.0
pandas==2.2.2
parse==1.20.2
PCSE==6.0.6
plotly==5.24.1
pmdarima==2.0.4
pvlib==0.11.0
pydantic==2.9.2
pydantic_settings==2.5.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
scipy==1.14.1
statsmodels==0.14.2
toga==0.4.6
util==0.0.1
